requirements.txt
confluent-kafka
requests
json5
pyspark
psycopg2-binary